+++ RELEASES.md
@@ -5,6 +5,13 @@ modification, are permitted in any medium without royalty provided the
 copyright notice and this notice are preserved. This file is offered
 as-is, without any warranty. -->
 
+Version 1.4.1 (unreleased)
+==========================
+
+  * For the experimental feature [`num-traits`][feat-exp-1-4], some
+    missing supertraits were added to
+    [`FixedOptionalFeatures`][tfof-1-4].
+
 Version 1.4.0 (2020-10-22)
 ==========================
 
@@ -45,6 +52,7 @@ Version 1.4.0 (2020-10-22)
 [nt-0-2-tp]: https://docs.rs/num-traits/^0.2/num_traits/cast/trait.ToPrimitive.html
 [nt-0-2-unsigned]: https://docs.rs/num-traits/^0.2/num_traits/sign/trait.Unsigned.html
 [tf-1-4]: https://docs.rs/fixed/~1.4/fixed/traits/trait.Fixed.html
+[tfof-1-4]: https://docs.rs/fixed/~1.4/fixed/traits/trait.FixedOptionalFeatures.html
 [w-1-4]: https://docs.rs/fixed/~1.4/fixed/struct.Wrapping.html
 
 Version 1.3.0 (2020-10-15)